# docs/swagger.yaml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: "/api"
info:
  version: "2.0.0"
  title: Jomon API
  description: >-
    JomonのAPIです。
...
paths:
...
  /admins:
    get:
      operationId: getAdmins
      description: adminユーザーの一覧を返す。管理者権限が必要。
      tags:
        - Admins
      responses:
        "200":
          description: 取得に成功した場合。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserID"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
    post:
      operationId: postAdmins
      description: adminユーザーを追加する。管理者権限が必要。
      tags:
        - Admins
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/UserID"
      responses:
        "200":
          description: 追加に成功した場合。
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
    delete:
      operationId: deleteAdmins
      description: adminユーザーを削除する。管理者権限が必要。
      tags:
        - Admins
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/UserID"
      responses:
        "200":
          description: 削除できた場合。
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
...
  /groups:
    get:
      operationId: getGroups
      description: グループの一覧を返す。
      tags:
        - Groups
      responses:
        "200":
          description: 取得に成功した場合。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: postGroup
      description: グループを追加する。管理者権限が必要。
      tags:
        - Groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostGroup"
      responses:
        "200":
          description: 追加に成功した場合。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
  /groups/{groupID}:
    parameters:
      - name: groupID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getGroupDetail
      description: グループの情報を返す。
      tags:
        - Groups
      responses:
        "200":
          description: 取得に成功した場合。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
        "404":
          $ref: "#/components/responses/404"
    put:
      operationId: putGroupDetail
      description: グループの情報を変更する。管理者権限またはグループオーナー権限が必要。
      tags:
        - Groups
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutGroup"
      responses:
        "200":
          description: 変更に成功した場合。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/400"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
    delete:
      operationId: deleteGroup
      description: グループを削除する。管理者権限またはグループオーナー権限が必要。
      tags:
        - Groups
      responses:
        "200":
          description: 削除に成功した場合
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
...
components:
  schemas:
...
    UserID:
      type: string
      format: uuid
...
    Group:
      type: object
      nullable: true
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "SysAd"
        description:
          type: string
          example: "SysAd班"
        budget:
          type: integer
          nullable: true
          example: 250000
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - budget
        - created_at
        - updated_at
    GroupDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "SysAd"
        description:
          type: string
          example: "SysAd班"
        budget:
          type: integer
          nullable: true
          example: 250000
        members:
          type: array
          items:
            $ref: "#/components/schemas/UserID"
        owners:
          type: array
          items:
            $ref: "#/components/schemas/UserID"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - budget
        - members
        - owners
        - created_at
        - updated_at
    PostGroup:
      type: object
      properties:
        name:
          type: string
          example: "SysAd"
        description:
          type: string
          example: "SysAd班"
        budget:
          type: integer
          nullable: true
          example: 250000
      required:
        - name
        - description
        - budget
    PutGroup:
      type: object
      properties:
        name:
          type: string
          example: "SysAd"
        description:
          type: string
          example: "SysAd班"
        budget:
          type: integer
          nullable: true
          example: 250000
      required:
        - name
        - description
        - budget
...
  responses:
...
    400:
      description: 不正なリクエスト。
    403:
      description: 編集権限がない人による操作。
    404:
      description: 指定したリソースは存在しない。
...